# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the VyoLayer API documentation.",
    "title": "VyoLayer API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:6999",
  "basePath": "/api/v1",
  "paths": {
...
    "/organizations/{orgId}/projects/{projectId}/api-keys": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["API Keys"],
        "summary": "List all API keys for a project",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dto.ApiKeyDTO"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["API Keys"],
        "summary": "Generate a new API key",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "description": "API key name",
            "name": "name",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "API key mode",
            "name": "mode",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/dto.ApiKeyCreatedDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/{orgId}/projects/{projectId}/api-keys/{apiKeyId}": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["API Keys"],
        "summary": "Get a specific API key by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "API key ID",
            "name": "apiKeyId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.ApiKeyDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/{orgId}/projects/{projectId}/api-keys/{apiKeyId}/revoke": {
      "post": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["API Keys"],
        "summary": "Revoke an API key",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Project ID",
            "name": "projectId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "API key ID",
            "name": "apiKeyId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dto.ApiKeyCreatedDTO": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "createdBy": {
          "type": "string",
          "example": "user_550e8400-e29b-41d4-a716-446655440000"
        },
        "expiresAt": {
          "type": "string",
          "example": "2024-01-01T00:00:00Z"
        },
        "id": {
          "type": "string",
          "example": "api_key_550e8400-e29b-41d4-a716-446655440000"
        },
        "isRevoked": {
          "type": "boolean",
          "example": false
        },
        "keyPrefix": {
          "type": "string",
          "example": "wl_live_ab3f1234"
        },
        "lastUsedAt": {
          "type": "string"
        },
        "mode": {
          "type": "string",
          "example": "live"
        },
        "name": {
          "type": "string",
          "example": "My API Key"
        },
        "projectId": {
          "type": "string",
          "example": "project_550e8400-e29b-41d4-a716-446655440000"
        },
        "rateLimit": {
          "type": "integer",
          "example": 60
        },
        "rawKey": {
          "type": "string",
          "example": "wl_live_a1b2c3d4e5f6..."
        },
        "requestLimit": {
          "type": "integer",
          "example": 1000
        },
        "revokedAt": {
          "type": "string"
        }
      }
    },
    "dto.ApiKeyDTO": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "createdBy": {
          "type": "string",
          "example": "user_550e8400-e29b-41d4-a716-446655440000"
        },
        "expiresAt": {
          "type": "string",
          "example": "2024-01-01T00:00:00Z"
        },
        "id": {
          "type": "string",
          "example": "api_key_550e8400-e29b-41d4-a716-446655440000"
        },
        "isRevoked": {
          "type": "boolean",
          "example": false
        },
        "keyPrefix": {
          "type": "string",
          "example": "wl_live_ab3f1234"
        },
        "lastUsedAt": {
          "type": "string"
        },
        "mode": {
          "type": "string",
          "example": "live"
        },
        "name": {
          "type": "string",
          "example": "My API Key"
        },
        "projectId": {
          "type": "string",
          "example": "project_550e8400-e29b-41d4-a716-446655440000"
        },
        "rateLimit": {
          "type": "integer",
          "example": 60
        },
        "requestLimit": {
          "type": "integer",
          "example": 1000
        },
        "revokedAt": {
          "type": "string"
        }
      }
    },
...
    "response.ErrorDetail": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "ERR_VALIDATION_FAILED"
        },
        "details": {},
        "message": {
          "type": "string",
          "example": "Validation failed"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": true
        },
        "requestId": {
          "type": "string",
          "example": "req-123456"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-02-08T18:30:00Z"
        }
      }
    },
    "response.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/response.ErrorDetail"
        },
        "statusCode": {
          "type": "integer",
          "example": 400
        },
        "success": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "response.Meta": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer",
          "example": 10
        },
        "page": {
          "type": "integer",
          "example": 1
        },
        "requestId": {
          "type": "string",
          "example": "req-123456"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-02-08T18:30:00Z"
        },
        "total": {
          "type": "integer",
          "example": 100
        },
        "totalPages": {
          "type": "integer",
          "example": 10
        }
      }
    },
    "response.SuccessResponse": {
      "type": "object",
      "properties": {
        "data": {},
        "message": {
          "type": "string",
          "example": "Operation successful"
        },
        "meta": {
          "$ref": "#/definitions/response.Meta"
        },
        "statusCode": {
          "type": "integer",
          "example": 200
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    }
  }
}